fe: updated navigation bar stories
